API reference

Index

HiddenMarkovModels.DiscreteMarkovChain

HiddenMarkovModels.DiscreteMarkovChainPrior

HiddenMarkovModels.DiscreteMarkovChainStats

HiddenMarkovModels.HMM

HiddenMarkovModels.HMMPrior

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.HiddenMarkovModelPrior

HiddenMarkovModels.baum_welch

HiddenMarkovModels.baum_welch_multiple_sequences

HiddenMarkovModels.baum_welch_multiple_sequences_log

HiddenMarkovModels.forward_backward!

HiddenMarkovModels.forward_backward_log!

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.nb_states

HiddenMarkovModels.randprobvec

HiddenMarkovModels.randtransmat

HiddenMarkovModels.stationary_distribution

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.uniformprobvec

HiddenMarkovModels.uniformtransmat

HiddenMarkovModels.update_obs_density!

HiddenMarkovModels.update_obs_logdensity!

Full docs

HiddenMarkovModels.DiscreteMarkovChain — Type

DiscreteMarkovChain

Discrete-time Markov chain with finite state space.

Fields

π0::AbstractVector: initial state distribution.

P::AbstractMatrix: state transition matrix.

HiddenMarkovModels.DiscreteMarkovChainPrior — Type

DiscreteMarkovChainPrior

Define a Dirichlet prior on the initial distribution and on the transition matrix of a DiscreteMarkovChain.

Fields

π0_α::AbstractVector: Dirichlet parameter for the initial distribution

P_α::AbstractMatrix: Dirichlet parameters for the transition matrix

HiddenMarkovModels.DiscreteMarkovChainStats — Type

DiscreteMarkovChainStats

Store sufficient statistics for the likelihood of a DiscreteMarkovChain sample.

Fields

initialization_count::AbstractVector: count initializations in each state

transition_count::AbstractMatrix: count transitions between each pair of states

HiddenMarkovModels.HMM — Type

HMM

Alias for HiddenMarkovModel.

HiddenMarkovModels.HMMPrior — Type

HMMPrior

Alias for HiddenMarkovModelPrior.

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{T,E}

Hidden Markov Model with arbitrary transition model (must be a discrete Markov chain) and emission distributions.

Fields

transitions::T: state evolution process.

emissions::Vector{E}: one emission distribution per state.

HiddenMarkovModels.HiddenMarkovModelPrior — Type

HiddenMarkovModelPrior{TP,EP}

Prior for a HiddenMarkovModel.

Fields

transitions_prior::T: prior on the transition structure.

emissions_prior::Vector{E}: one prior per state emission distribution.

Base.rand — Method

rand([rng,] mc::DiscreteMarkovChain, T)

Simulate mc during T time steps.

Base.rand — Method

rand([rng,] prior::DiscreteMarkovChainPrior)

Sample a DiscreteMarkovChain from prior.

Base.rand — Method

rand([rng,] hmm::HMM, T)

Sample a sequence of states of length T and the associated sequence of observations.

DensityInterface.logdensityof — Method

logdensityof(mc::DiscreteMarkovChain, x::AbstractVector)

Compute the log-likelihood of the sequence x of states for the chain mc.

DensityInterface.logdensityof — Method

logdensityof(prior::DiscreteMarkovChainPrior, mc::DiscreteMarkovChain)

Compute the log-likelihood of the chain mc with respect to a prior.

HiddenMarkovModels.baum_welch — Method

baum_welch(hmm_init, observation_sequence; log, plot)

Run the Baum-Welch algorithm to estimate a HiddenMarkovModel of the same type as hmm_init, based on a single observation_sequence.

HiddenMarkovModels.baum_welch_multiple_sequences — Method

baum_welch_multiple_sequences(hmm_init, observation_sequences; iterations)

Run the Baum-Welch algorithm to estimate a HiddenMarkovModel of the same type as hmm_init, based on several observation_sequences.

HiddenMarkovModels.baum_welch_multiple_sequences_log — Method

baum_welch_multiple_sequences_log(hmm_init, observation_sequences; iterations)

Same as baum_welch_multiple_sequences but with logarithmic computations for numerical stability.

HiddenMarkovModels.forward_backward! — Method

forward_backward!(α, β, c, γ, ξ, obs_density, hmm)

Apply the forward-backward algorithm in-place to compute sufficient statistics.

HiddenMarkovModels.forward_backward_log! — Method

forward_backward!(logα, logβ, logγ, logξ, obs_logdensity, hmm)

Apply the logarithmic forward-backward algorithm in-place to compute sufficient statistics.

HiddenMarkovModels.initial_distribution — Method

initial_distribution(mc::DiscreteMarkovChain)

Return the vector of initial state probabilities of mc.

HiddenMarkovModels.nb_states — Method

nb_states(mc::DiscreteMarkovChain)

Return the number of states of mc.

HiddenMarkovModels.randprobvec — Method

randprobvec(n)

Return a random probability distribution vector of size n.

HiddenMarkovModels.randtransmat — Method

randtransmat(n)

Return a stochastic matrix of size n with random transition probability distributions.

HiddenMarkovModels.stationary_distribution — Method

stationary_distribution(mc::DiscreteMarkovChain)

Compute the equilibrium distribution of mc using its eigendecomposition.

HiddenMarkovModels.transition_matrix — Method

transition_matrix(mc::DiscreteMarkovChain)

Return the matrix of transition probabilities of mc.

HiddenMarkovModels.uniformprobvec — Method

uniformprobvec(n)

Return a uniform probability distribution vector of size n.

HiddenMarkovModels.uniformtransmat — Method

uniformtransmat(n)

Return a stochastic matrix of size n with uniform transition probability distributions.

HiddenMarkovModels.update_obs_density! — Method

update_obs_density!(obs_density, observations, hmm)

Set obs_density[t, s] to the likelihood of hmm emitting observation[t] if it were in state s.

HiddenMarkovModels.update_obs_logdensity! — Method

update_obs_logdensity!(obs_density, observations, hmm)

Set obs_logdensity[t, s] to the log-likelihood of hmm emitting observation[t] if it were in state s.