API reference

Index

HiddenMarkovModels.DiscreteMarkovChain

HiddenMarkovModels.DiscreteMarkovChainPrior

HiddenMarkovModels.DiscreteMarkovChainStats

HiddenMarkovModels.HMM

HiddenMarkovModels.HMMPrior

HiddenMarkovModels.HiddenMarkovModel

HiddenMarkovModels.HiddenMarkovModelPrior

HiddenMarkovModels.History

HiddenMarkovModels.MultivariatePoissonProcess

HiddenMarkovModels.baum_welch

HiddenMarkovModels.baum_welch_multiple_sequences

HiddenMarkovModels.duration

HiddenMarkovModels.flat_prior

HiddenMarkovModels.forward_backward!

HiddenMarkovModels.has_events

HiddenMarkovModels.initial_distribution

HiddenMarkovModels.max_mark

HiddenMarkovModels.max_time

HiddenMarkovModels.min_time

HiddenMarkovModels.nb_events

HiddenMarkovModels.nb_states

HiddenMarkovModels.rand_prob_vec

HiddenMarkovModels.rand_trans_mat

HiddenMarkovModels.split_into_chunks

HiddenMarkovModels.stationary_distribution

HiddenMarkovModels.time_change

HiddenMarkovModels.transition_matrix

HiddenMarkovModels.uniform_prob_vec

HiddenMarkovModels.uniform_trans_mat

HiddenMarkovModels.update_obs_density!

Full docs

HiddenMarkovModels.DiscreteMarkovChain — Type

DiscreteMarkovChain

Discrete-time Markov chain with finite state space.

Fields

p0::AbstractVector: initial state distribution.

P::AbstractMatrix: state transition matrix.

HiddenMarkovModels.DiscreteMarkovChainPrior — Type

DiscreteMarkovChainPrior

Define a Dirichlet prior on the initial distribution and on the transition matrix of a DiscreteMarkovChain.

Fields

p0_α::AbstractVector: Dirichlet parameter for the initial distribution

P_α::AbstractMatrix: Dirichlet parameters for the transition matrix

HiddenMarkovModels.DiscreteMarkovChainStats — Type

DiscreteMarkovChainStats

Store sufficient statistics for the likelihood of a DiscreteMarkovChain sample.

Fields

initialization_count::AbstractVector: count initializations in each state

transition_count::AbstractMatrix: count transitions between each pair of states

HiddenMarkovModels.HMM — Type

HMM

Alias for HiddenMarkovModel.

HiddenMarkovModels.HMMPrior — Type

HMMPrior

Alias for HiddenMarkovModelPrior.

HiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{Tr,Em}

Hidden Markov Model with arbitrary transition model (must be a discrete Markov chain) and emission distributions.

Fields

transitions::Tr: state evolution process.

emissions::Vector{Em}: one emission distribution per state.

HiddenMarkovModels.HiddenMarkovModelPrior — Type

HiddenMarkovModelPrior{TrP,EmP}

Prior for a HiddenMarkovModel.

Fields

transitions_prior::TrP: prior on the transition structure.

emissions_prior::Vector{EmP}: one prior per state emission distribution.

HiddenMarkovModels.History — Type

History{M,T<:Real}

Linear event histories with marks of type M and locations of real type T.

Fields

times::Vector{T}: vector of event times

marks::Vector{M}: vector of event marks

tmin::T: start time

tmax::T: end time

HiddenMarkovModels.MultivariatePoissonProcess — Type

MultivariatePoissonProcess{R}

Multivariate homogeneous temporal Poisson process.

Fields

λ::Vector{R}: event rates.

Base.append! — Method

append!(h1::History, h2::History)

Add all the events of h2 at the end of h1.

Base.push! — Method

push!(h::History, t, m)

Add event (t, m) at the end of history h.

Base.rand — Method

rand([rng,] mc::DiscreteMarkovChain, T)

Simulate mc during T time steps.

Base.rand — Method

rand([rng,] prior::DiscreteMarkovChainPrior)

Sample a DiscreteMarkovChain from prior.

Base.rand — Method

rand([rng,] hmm::HMM, T)

Sample a sequence of states of length T and the associated sequence of observations.

DensityInterface.logdensityof — Method

logdensityof(mc::DiscreteMarkovChain, x::AbstractVector)

Compute the log-likelihood of the sequence x of states for the chain mc.

DensityInterface.logdensityof — Method

logdensityof(prior::DiscreteMarkovChainPrior, mc::DiscreteMarkovChain)

Compute the log-likelihood of the chain mc with respect to a prior.

HiddenMarkovModels.baum_welch — Method

baum_welch(hmm_init, obs_sequence)

Same as baum_welch_multiple_sequences but with a single sequence.

HiddenMarkovModels.baum_welch_multiple_sequences — Method

baum_welch_multiple_sequences(hmm_init, obs_sequences)

Run the Baum-Welch algorithm to estimate a HiddenMarkovModel of the same type as hmm_init, based on several observation sequences.

HiddenMarkovModels.duration — Method

duration(h::History)

Compute the difference h.tmax - h.tmin.

HiddenMarkovModels.flat_prior — Method

zero_prior(mc::DiscreteMarkovChain)

Build a flat prior, for which MAP is equivalent to MLE.

HiddenMarkovModels.forward_backward! — Method

forward_backward!(α, β, γ, ξ, α_sum_inv, γ_sum_inv, ξ_sum_inv, hmm, obs_density)

Apply the forward-backward algorithm in-place to update sufficient statistics.

HiddenMarkovModels.has_events — Function

has_events(h::History, tmin=-Inf, tmax=Inf)

Check the presence of events in h during the interval [tmin, tmax).

HiddenMarkovModels.initial_distribution — Method

initial_distribution(mc::DiscreteMarkovChain)

Return the vector of initial state probabilities of mc.

HiddenMarkovModels.max_mark — Method

max_mark(h::History)

Compute the highest mark, provided they can be ordered.

HiddenMarkovModels.max_time — Method

max_time(h)

Return the end time of h (not the same as the last event time).

HiddenMarkovModels.min_time — Method

min_time(h)

Return the starting time of h (not the same as the first event time).

HiddenMarkovModels.nb_events — Function

nb_events(h::History, tmin=-Inf, tmax=Inf)

Count events in h during the interval [tmin, tmax).

HiddenMarkovModels.nb_states — Method

nb_states(mc::DiscreteMarkovChain)

Return the number of states of mc.

HiddenMarkovModels.rand_prob_vec — Method

rand_prob_vec(n)

Return a random probability distribution vector of size n.

HiddenMarkovModels.rand_trans_mat — Method

rand_trans_mat(n)

Return a stochastic matrix of size n with random transition probability distributions.

HiddenMarkovModels.split_into_chunks — Method

split_into_chunks(h, chunk_duration)

Split h into a vector of consecutive histories with individual duration chunk_duration.

HiddenMarkovModels.stationary_distribution — Method

stationary_distribution(mc::DiscreteMarkovChain)

Compute the equilibrium distribution of mc using its eigendecomposition.

HiddenMarkovModels.time_change — Method

time_change(h, Λ)

Apply the time rescaling t -> Λ(t) to history h.

HiddenMarkovModels.transition_matrix — Method

transition_matrix(mc::DiscreteMarkovChain)

Return the matrix of transition probabilities of mc.

HiddenMarkovModels.uniform_prob_vec — Method

uniform_prob_vec(n)

Return a uniform probability distribution vector of size n.

HiddenMarkovModels.uniform_trans_mat — Method

uniform_trans_mat(n)

Return a stochastic matrix of size n with uniform transition probability distributions.

HiddenMarkovModels.update_obs_density! — Method

update_obs_density!(obs_density, hmm, obs_sequence)

Set obs_density[s, t] to the likelihood of hmm emitting obs_sequence[t] if it were in state s.