API reference

ControlledHiddenMarkovModels.CHMMs — Module

CHMMs

Alias for ControlledHiddenMarkovModels.

ControlledHiddenMarkovModels.ControlledHiddenMarkovModels — Module

ControlledHiddenMarkovModels

A package for Hidden Markov Models with exogenous control variables.

Types

ControlledHiddenMarkovModels.AbstractControlledHMM — Type

AbstractControlledHMM

Alias for AbstractControlledHiddenMarkovModel.

ControlledHiddenMarkovModels.AbstractControlledHiddenMarkovModel — Type

AbstractControlledHiddenMarkovModel

Interface for Hidden Markov Models with arbitrary emissions and exogenous control variables.

Required methods

nb_states(hmm, par)

initial_distribution(hmm, par)

transition_matrix(hmm, control, par)

transition_matrix!(P, hmm, control, par)

emission_parameters(hmm, control, par)

emission_parameters!(θ, hmm, control, par)

emission_distribution(hmm, s, θ)

Optional methods

log_initial_distribution(hmm, par)

log_transition_matrix(hmm, control, par)

log_transition_matrix!(logP, hmm, control, par)

Compatible with

rand(rng, hmm, control_sequence, par)

logdensityof(hmm, obs_sequence, control_sequence, par; safe)

infer_current_state(hmm, obs_sequence, control_sequence, par; safe)

ControlledHiddenMarkovModels.AbstractHMM — Type

AbstractHMM

Alias for AbstractHiddenMarkovModel.

ControlledHiddenMarkovModels.AbstractHiddenMarkovModel — Type

AbstractHiddenMarkovModel

Interface for Hidden Markov Models with arbitrary emissions.

Required methods

nb_states(hmm, par)

initial_distribution(hmm, par)

transition_matrix(hmm, par)

emission_distribution(hmm, s, par)

Optional methods

log_initial_distribution(hmm, par)

log_transition_matrix(hmm, par)

Compatible with

rand(rng, hmm, T, par)

logdensityof(hmm, obs_sequence, par; safe)

infer_current_state(hmm, obs_sequence, par; safe)

ControlledHiddenMarkovModels.ForwardBackwardStorage — Type

ForwardBackwardStorage{R}

Storage for the forward-backward algorithm applied to multiple sequences.

Fields

α::Vector{Matrix{R}}

c::Vector{Vector{R}}

β::Vector{Matrix{R}}

bβ::Vector{Matrix{R}}

γ::Vector{Matrix{R}}

ξ::Vector{Array{R,3}}

ControlledHiddenMarkovModels.HMM — Type

HMM

Alias for HiddenMarkovModel.

ControlledHiddenMarkovModels.HiddenMarkovModel — Type

HiddenMarkovModel{R1,R2,D}

Concrete subtype of AbstractHMM which stores the state and observation parameters directly.

Fields

p0::Vector{R1}

P::Matrix{R2}

emissions::Vector{D}

Compatible with

baum_welch_multiple_sequences(obs_sequences, hmm_init, par; kwargs...)

baum_welch(obs_sequence, hmm_init, par; kwargs...)

ControlledHiddenMarkovModels.LogForwardBackwardStorage — Type

LogForwardBackwardStorage{R}

Storage for the forward-backward algorithm in log scale applied to multiple sequences.

Fields

logα::Vector{Matrix{R}}

logβ::Vector{Matrix{R}}

logγ::Vector{Matrix{R}}

logξ::Vector{Array{R,3}}

Methods

Base.rand — Function

rand(rng, hmm::AbstractHMM, T, par)

Sample a trajectory of length T from hmm with parameters par.

Returns a couple (state_sequence, obs_sequence).

Base.rand — Method

rand(rng, hmm::AbstractControlledHMM, control_sequence, par)

Sample a trajectory from hmm with controls control_sequence and parameters par.

Returns a couple (state_sequence, obs_sequence).

ControlledHiddenMarkovModels.backward_doublelog! — Method

backward_doublelog!(logβ, obs_logdensity, hmm::AbstractHMM, par)

Perform a backward pass fully in log scale by mutating logβ.

ControlledHiddenMarkovModels.backward_log! — Method

backward_log!(β, bβ, c, obs_logdensity, hmm::AbstractHMM, par)

Perform a backward pass by mutating β, bβ and c.

ControlledHiddenMarkovModels.backward_nolog! — Method

backward_nolog!(β, bβ, c, obs_density, hmm::AbstractHMM, par)

Perform a backward pass by mutating β, bβ and c.

ControlledHiddenMarkovModels.baum_welch_doublelog — Function

baum_welch_doublelog(obs_sequences, hmm_init::HMM[, par; maxiter, tol])

Apply the Baum-Welch algorithm fully in log scale on multiple observation sequences, starting from an initial HMM hmm_init with parameters par (not modifed).

ControlledHiddenMarkovModels.baum_welch_log — Function

baum_welch_log(obs_sequences, hmm_init::HMM[, par; maxiter, tol])

Apply the Baum-Welch algorithm partly in log scale on multiple observation sequences, starting from an initial HMM hmm_init with parameters par (not modifed).

ControlledHiddenMarkovModels.baum_welch_nolog — Function

baum_welch_nolog(obs_sequences, hmm_init::HMM[, par; maxiter, tol])

Apply the Baum-Welch algorithm on multiple observation sequences, starting from an initial HMM hmm_init with parameters par (not modifed).

ControlledHiddenMarkovModels.emission_distribution — Method

emission_distribution(hmm::AbstractControlledHMM, s, θ)

Compute the emission distribution in state s for hmm with emission parameters θ. Note that θ was computed using emission_parameters(hmm, control, par).

The object returned must be sampleable and implement DensityInterface.jl.

ControlledHiddenMarkovModels.emission_distribution — Method

emission_distribution(hmm::AbstractHMM, s, par)

Compute the emission distribution in state s for hmm with parameters par.

The object returned must be sampleable and implement DensityInterface.jl.

ControlledHiddenMarkovModels.emission_parameters! — Method

emission_parameters!(θ, hmm::AbstractControlledHMM, control, par)

Update θ with the parameters of all emission distributions for hmm with control control and parameters par.

ControlledHiddenMarkovModels.emission_parameters — Method

emission_parameters(hmm::AbstractControlledHMM, control, par)

Compute the parameters of all emission distributions for hmm with control control and parameters par.

ControlledHiddenMarkovModels.emission_type — Method

emission_type(::Type{<:HMM})

Return the type of an emission distribution object.

ControlledHiddenMarkovModels.fit_mle_from_multiple_sequences — Method

fit_mle_from_multiple_sequences(::Type{D}, xs, ws)

Fit a distribution of type D based on multiple sequences of observations xs associated with multiple sequences of weights ws.

Must accept arbitrary iterables for xs and ws.

ControlledHiddenMarkovModels.fit_mle_from_single_sequence — Method

fit_mle_from_single_sequence(::Type{D}, x, w)

Fit a distribution of type D based on a single sequence of observations x associated with a single sequence of weights w.

Defaults to Distributions.fit_mle, with a special case for vectors of vectors (because Distributions.fit_mle accepts matrices instead). Users are free to override this default for concrete distributions.

ControlledHiddenMarkovModels.forward_backward_doublelog! — Method

forward_backward_doublelog!(logα, logβ, logγ, logξ, obs_logdensity, hmm::AbstractHMM, par)

Apply the full forward-backward algorithm fully in log scale by mutating logα, logβ, logγ and logξ.

ControlledHiddenMarkovModels.forward_backward_log! — Method

forward_backward_nolog!(α, c, β, bβ, γ, ξ, obs_logdensity, hmm::AbstractHMM, par)

Apply the full forward-backward algorithm partly in log scale by mutating α, c, β, bβ, γ and ξ.

ControlledHiddenMarkovModels.forward_backward_nolog! — Method

forward_backward_nolog!(α, c, β, bβ, γ, ξ, obs_density, hmm::AbstractHMM, par)

Apply the full forward-backward algorithm by mutating α, c, β, bβ, γ and ξ.

ControlledHiddenMarkovModels.forward_doublelog! — Method

forward_doublelog!(logα, obs_logdensity, hmm::AbstractHMM, par)

Perform a forward pass fully in log scale by mutating logα.

ControlledHiddenMarkovModels.forward_log! — Method

forward_log!(α, c, obs_logdensity, hmm::AbstractHMM, par)

Perform a forward pass partly in log scale by mutating α and c.

ControlledHiddenMarkovModels.forward_nolog! — Method

forward_nolog!(α, c, obs_density, hmm::AbstractHMM, par)

Perform a forward pass by mutating α and c.

ControlledHiddenMarkovModels.infer_current_state — Function

infer_current_state(hmm::AbstractHMM, obs_sequence, par; safe)

Infer the posterior distribution of the current state given obs_sequence for hmm with parameters par.

ControlledHiddenMarkovModels.infer_current_state — Function

infer_current_state(
    hmm::AbstractControlledHMM, obs_sequence, control_sequence, par; safe
)

Infer the posterior distribution of the current state given obs_sequence for hmm with controls control_sequence and parameters par.

ControlledHiddenMarkovModels.initial_distribution — Method

initial_distribution(hmm::AbstractControlledHMM, par)

Return the vector of initial state probabilities for hmm with parameters par.

ControlledHiddenMarkovModels.initial_distribution — Method

initial_distribution(hmm::AbstractHMM, par)

Compute the vector of initial state probabilities for hmm with parameters par.

ControlledHiddenMarkovModels.initialize_obs_density — Method

initialize_obs_density(obs_sequence, hmm, par)

Create a new observation density matrix and apply update_obs_density!.

ControlledHiddenMarkovModels.initialize_obs_logdensity — Method

initialize_obs_logdensity(obs_sequence, hmm, par)

Create a new observation log-density matrix and apply update_obs_logdensity!.

ControlledHiddenMarkovModels.is_prob_vec — Method

is_prob_vec(p; atol)

Check if p is a probability distribution vector.

ControlledHiddenMarkovModels.iszero_safe — Method

iszero_safe(x::R)

Check if a number x is zero by comparing its inverse with typemax(R).

This is useful in the following case:

julia> x = 1e-320
1.0e-320

julia> iszero(x)
false

julia> inv(x)
Inf

ControlledHiddenMarkovModels.light_forward_doublelog — Method

light_forward_doublelog(obs_sequence, hmm::AbstractHMM, par)

Perform a lightweight forward pass fully in log scale with minimal storage requirements.

ControlledHiddenMarkovModels.light_forward_doublelog — Method

light_forward_doublelog(obs_sequence, control_sequence, hmm::AbstractControlledHMM, par)

Perform a lightweight forward pass fully in log scale with minimal storage requirements.

ControlledHiddenMarkovModels.light_forward_log — Method

light_forward_log(obs_sequence, hmm::AbstractHMM, par)

Perform a lightweight forward pass partly in log scale with minimal storage requirements.

ControlledHiddenMarkovModels.light_forward_log — Method

light_forward_log(obs_sequence, control_sequence, hmm::AbstractControlledHMM, par)

Perform a lightweight forward pass partly in log scale with minimal storage requirements.

ControlledHiddenMarkovModels.light_forward_nolog — Method

light_forward_nolog(obs_sequence, hmm::AbstractHMM, par)

Perform a lightweight forward pass with minimal storage requirements.

ControlledHiddenMarkovModels.light_forward_nolog — Method

light_forward_nolog(obs_sequence, control_sequence, hmm::AbstractControlledHMM, par)

Perform a lightweight forward pass with minimal storage requirements.

ControlledHiddenMarkovModels.log_initial_distribution — Method

log_initial_distribution(hmm::AbstractControlledHMM, par)

Return the vector of initial state probabilities in log scale for hmm with parameters par.

ControlledHiddenMarkovModels.log_initial_distribution — Method

log_initial_distribution(hmm::AbstractHMM, par)

Compute the vector of initial state probabilities in log scale for hmm with parameters par.

ControlledHiddenMarkovModels.log_transition_matrix! — Method

log_transition_matrix!(logP, hmm::AbstractControlledHMM, control, par)

Update logP with the state transition matrix in log scale for hmm with control control and parameters par.

ControlledHiddenMarkovModels.log_transition_matrix — Method

log_transition_matrix(hmm::AbstractControlledHMM, control, par)

Compute the state transition matrix in log scale for hmm with control control and parameters par.

ControlledHiddenMarkovModels.log_transition_matrix — Method

log_transition_matrix(hmm::AbstractHMM, par)

Compute the state transition matrix in log scale for hmm with parameters par.

ControlledHiddenMarkovModels.logsumexp — Method

logsumexp(a)

Use logsumexp_stream to compute the logsumexp of an iterable a.

ControlledHiddenMarkovModels.logsumexp_stream — Method

logsumexp_stream(::Type{T}, a)

Compute the logsumexp function in a single pass for an iterable a with elements of type T.

Source: https://www.nowozin.net/sebastian/blog/streaming-log-sum-exp-computation.html

ControlledHiddenMarkovModels.make_log_prob_vec! — Method

make_log_prob_vec!(logp)

Shift logp so that exp.(logp) becomes a probability distribution vector.

ControlledHiddenMarkovModels.make_log_trans_mat! — Method

make_log_trans_mat!(logP)

Scale logP so that exp.(logP) becomes a transition (stochastic) matrix.

ControlledHiddenMarkovModels.make_prob_vec! — Method

make_prob_vec!(p)

Scale p into a probability distribution vector.

ControlledHiddenMarkovModels.make_trans_mat! — Method

make_trans_mat!(P)

Scale P into a transition (stochastic) matrix.

ControlledHiddenMarkovModels.nb_states — Method

nb_states(hmm::AbstractControlledHMM, par)

Return the number of states for hmm with parameters par.

ControlledHiddenMarkovModels.nb_states — Method

nb_states(hmm::AbstractHMM, par)

Compute the number of states for hmm with parameters par.

ControlledHiddenMarkovModels.rand_prob_vec — Method

rand_prob_vec(rng, n)

Return a random probability distribution vector of size n.

ControlledHiddenMarkovModels.rand_trans_mat — Method

rand_trans_mat(rng, n)

Return a transition (stochastic) matrix of size n with random transition probability distributions.

ControlledHiddenMarkovModels.transition_matrix! — Method

transition_matrix!(P, hmm::AbstractControlledHMM, control, par)

Update P with the state transition matrix for hmm with control control and parameters par.

ControlledHiddenMarkovModels.transition_matrix — Method

transition_matrix(hmm::AbstractControlledHMM, control, par)

Compute the state transition matrix for hmm with control control and parameters par.

ControlledHiddenMarkovModels.transition_matrix — Method

transition_matrix(hmm::AbstractHMM, par)

Compute the state transition matrix for hmm with parameters par.

ControlledHiddenMarkovModels.uniform_prob_vec — Method

uniform_prob_vec(n)

Return a uniform probability distribution vector of size n.

ControlledHiddenMarkovModels.uniform_trans_mat — Method

uniform_trans_mat(n)

Return a transition (stochastic) matrix of size n with uniform transition probability distributions.

ControlledHiddenMarkovModels.update_obs_density! — Method

update_obs_density!(obs_density, obs_sequence, hmm::AbstractHMM, par)

Update the values obs_density[s, t] using the emission density of hmm with parameters par applied to obs_sequence[t].

ControlledHiddenMarkovModels.update_obs_logdensity! — Method

update_obs_logdensity!(obs_logdensity, obs_sequence, hmm::AbstractHMM, par)

Update the values obs_logdensity[s, t] using the emission log-density of hmm with parameters par applied to obs_sequence[t].

DensityInterface.logdensityof — Function

logdensityof(hmm::AbstractHMM, obs_sequence, par; safe)

Compute the log likelihood of obs_sequence for hmm with parameters par.

DensityInterface.logdensityof — Function

logdensityof(hmm::AbstractControlledHMM, obs_sequence, control_sequence, par; safe)

Compute the log likelihood of obs_sequence for hmm with controls control_sequence and parameters par.

Index

ControlledHiddenMarkovModels.CHMMs

ControlledHiddenMarkovModels.ControlledHiddenMarkovModels

ControlledHiddenMarkovModels.AbstractControlledHMM

ControlledHiddenMarkovModels.AbstractControlledHiddenMarkovModel

ControlledHiddenMarkovModels.AbstractHMM

ControlledHiddenMarkovModels.AbstractHiddenMarkovModel

ControlledHiddenMarkovModels.ForwardBackwardStorage

ControlledHiddenMarkovModels.HMM

ControlledHiddenMarkovModels.HiddenMarkovModel

ControlledHiddenMarkovModels.LogForwardBackwardStorage

ControlledHiddenMarkovModels.backward_doublelog!

ControlledHiddenMarkovModels.backward_log!

ControlledHiddenMarkovModels.backward_nolog!

ControlledHiddenMarkovModels.baum_welch_doublelog

ControlledHiddenMarkovModels.baum_welch_log

ControlledHiddenMarkovModels.baum_welch_nolog

ControlledHiddenMarkovModels.emission_distribution

ControlledHiddenMarkovModels.emission_distribution

ControlledHiddenMarkovModels.emission_parameters

ControlledHiddenMarkovModels.emission_parameters!

ControlledHiddenMarkovModels.emission_type

ControlledHiddenMarkovModels.fit_mle_from_multiple_sequences

ControlledHiddenMarkovModels.fit_mle_from_single_sequence

ControlledHiddenMarkovModels.forward_backward_doublelog!

ControlledHiddenMarkovModels.forward_backward_log!

ControlledHiddenMarkovModels.forward_backward_nolog!

ControlledHiddenMarkovModels.forward_doublelog!

ControlledHiddenMarkovModels.forward_log!

ControlledHiddenMarkovModels.forward_nolog!

ControlledHiddenMarkovModels.infer_current_state

ControlledHiddenMarkovModels.infer_current_state

ControlledHiddenMarkovModels.initial_distribution

ControlledHiddenMarkovModels.initial_distribution

ControlledHiddenMarkovModels.initialize_obs_density

ControlledHiddenMarkovModels.initialize_obs_logdensity

ControlledHiddenMarkovModels.is_prob_vec

ControlledHiddenMarkovModels.iszero_safe

ControlledHiddenMarkovModels.light_forward_doublelog

ControlledHiddenMarkovModels.light_forward_doublelog

ControlledHiddenMarkovModels.light_forward_log

ControlledHiddenMarkovModels.light_forward_log

ControlledHiddenMarkovModels.light_forward_nolog

ControlledHiddenMarkovModels.light_forward_nolog

ControlledHiddenMarkovModels.log_initial_distribution

ControlledHiddenMarkovModels.log_initial_distribution

ControlledHiddenMarkovModels.log_transition_matrix

ControlledHiddenMarkovModels.log_transition_matrix

ControlledHiddenMarkovModels.log_transition_matrix!

ControlledHiddenMarkovModels.logsumexp

ControlledHiddenMarkovModels.logsumexp_stream

ControlledHiddenMarkovModels.make_log_prob_vec!

ControlledHiddenMarkovModels.make_log_trans_mat!

ControlledHiddenMarkovModels.make_prob_vec!

ControlledHiddenMarkovModels.make_trans_mat!

ControlledHiddenMarkovModels.nb_states

ControlledHiddenMarkovModels.nb_states

ControlledHiddenMarkovModels.rand_prob_vec

ControlledHiddenMarkovModels.rand_trans_mat

ControlledHiddenMarkovModels.transition_matrix

ControlledHiddenMarkovModels.transition_matrix

ControlledHiddenMarkovModels.transition_matrix!

ControlledHiddenMarkovModels.uniform_prob_vec

ControlledHiddenMarkovModels.uniform_trans_mat

ControlledHiddenMarkovModels.update_obs_density!

ControlledHiddenMarkovModels.update_obs_logdensity!